Basic docker setup
This provides a basic dataverse instance running in a docker for the
Linux CI build by utilizing https://github.com/IQSS/dataverse-docker.
While it may be possible to run it on Windows, too, at least the compose
file is not ready for that (posix paths, etc.)

The setup uses the provided default recipe and runs its `setup-users.sh`
script to have the demo accounts 'pete' and 'uma' and get their API
tokens. The tokens are then stored in the environment variables
`TEST_TOKEN_PETE` and `TEST_TOKEN_UMA` respectively for use with the
actual tests.