TST: Improve test setup
This introduces `datalad_dataverse.tests.API_TOKENS` and
`datalad_dataverse.tests.DATAVERSEURL`, both of which are taken from
environment variables. This allows the CI setup as well as developers to
run the tests, given the base URL is provided in env var
`TESTS_DATAVERSE_BASEURL` and user tokens in `TESTS_TOKEN_<USERNAME>`.

This allows to easily decorate tests with
`@skip_if(cond=not datalad_dataverse.tests.DATAVERSEURL)` to not execute
them without a such a setup.