Hook into the datalad-next credential system
There is not longer a need to pre-deploy a token via an env-var. The
special remote will no query for one automatically and also
store it in a local credential store on successful use. Subsequent
usage no longer requires a re-specification, because the credential
is rediscovered via a "realm" property.