Enable test execution against demo.dataverse.org
The change makes it possible to run all tests against this common
instance, after specifying a token. The standard user account token
handed out to any user of this instance is a sufficiently capable
admin token.

When no particular dataverse instance URL is specified, all tests
are performed against demo.dataverse.org. This substantially lowers
the barrier for running tests, at the cost of a much slower test
execution.

When no admin token for the tests is given, a prominent warning
with instructions is now given, rather than merely skipping most tests.

The environment variable scheme for configurating the test setup has
been standardized to the following setup:

- DATAVERSE_TEST_URL: Dataverse instance base URL, defaults to
  https://demo.dataverse.org
- DATAVERSE_TEST_BASECOLLECTION: defaults to 'demo' for demo.dataverse.org
  and to 'root' for any other instance URL
- DATAVERSE_TEST_APITOKEN_<USER>: tokens specification for any user

Variable names in the Python code have been changes to match the
environment variable names to prevent confusion and the need to
look up the mapping.