Run tests against demo.dataverse.org when no docker deployment is around
This is true to crippled FS runs, and appveyor's Mac and Win runs.

The necessary token is stored as a GitHub Action secret and in the
appveyor environment specification. The token has a lifetime of one
year and needs to be updated afterwards. @mih know the credentials
for the associated account on this instance.