Reimplement special remote
This essentially amounts to a rewrite of the special remote.
Core piece is to rely on tracking dataverse's database ID per key
instead of plain path matching via annex' SETSTATE/GETSTATE.

However, path matching is still partially required as a fallback:
1. Required: The special keys of the `datalad-annex` git remote helper
   need to be matched by path, since the throwaway repository using them
   can't have access to a git-annex branch delivering the ID (especially
   when cloning from datalad-annex URL)
2. In regular non-export mode, we store keys under their name. Matching
   them as a fallback has little potential to go wrong (someone would
   need to upload mismatching content under an annex keys' filename),
   while it would provide some means to recover from a messed up
   situation. Say, someone uploaded via `annex-copy` w/o pushing the
   annex branch. Or the annex branch was messed up somehow. With
   filename matching as fallback there would still be a way to recover
   the data w/o reuploading. So - for annex mode I think it's worth it
   as fallback.

This implementation should be faster as it saves a bunch of requests.
- `checkpresent(export)` used to send the same request every time. This
  is now cached and updated locally when we push changes
  (`self.files_*`) from within the same process.
- implemented `renameexport` to only change the metadata of an uploaded
  file when it was moved. W/O this method being implemented, annex would
  do a REMOVEEXPORT + TRANSFEREXPORT_STORE which is (a lot) more
  expensive.

Uploading a file with the same path (already present in the latest
version of the dataverse dataset) would lead to dataverse renaming the
file. This is now addressed by replacing the file instead under those
conditions.

Mangling directory names to solve the leading dot issue was failing when
dealing with a path whose parent was ".", since accessing its `.parent`
gives an empty tuple. Therefore exporting files at the root directory of
the dataset failed.

Implements RENAMEEXPORT. If not implemented by the special remote, annex
will perform REMOVE + RE-UPLOAD instead. While it works, it doesn't
scale well. However, implementation based on pyDataverse's NativeAPI is
somewhat weird, as it relies on `curl` rather than the `requests` lib,
which in turn means that availability of this method depends on
availability of `curl`.