Consolidate all prep into `prepare()`
Before basic things like the url, doi, or the API connection
were obtained on-demand. However, in practice all these things
need to be around all the time.

This is now done in `prepare()`. And given that even `initremote()`
needs to same setup, it is reused in there too.

I have removed the associated properties. In general they are not bad,
but for now not needed and can come back if that changes.

Ping #170