Rectify assumptions for file removal
This fix is needed after ripping out the special casing for XDLRA
keys. It intentionally only addresses the non-export case, to make
clear what is important for which mode.

This fix was developed by @christian-monch as part of
https://github.com/mih/datalad-dataverse/pull/1

It fixes the situation where a git-clone from dataverse cannot know
the fileId of a repository export (datalad-annex git remote helper)
that, by definition, needs to be packed up and uploaded _before_
a fileId can be known and registered in the repository that is
already finalized and uploaded.

In order to break this chicken-and-egg-problem, `_remove_file()` now
uniformly falls back on determining the fileId via path matching.

But see #189 and #188 for related aspects of this general issue.