Replace non-sensical call
`self.remove_from_filelist(re)`, where `re` is the Python regex module.

It is now passing `replace_id`, which is checked for `not None` directly
before. However, this is a bling fix. Clearly this part of the code is
not tested enough to be able to tell whether this is actually a fix.

Closes #195