Add test to document file removal via SWORD API
The only documented way of removing a file from a dataverse dataset is
via SWORD API. This requires a database id of the file (as opposed to a
fie name). We cannot remove a file, once it is part of a published
dataset version. Removal should, however, work if a file is not yet
published.

Document this in a test.