Add more testing to `test_remote`
So for it was mostly a smoke test. Now do checks on the targeted remote
file naming. No other test is doing that, and failure was only
indirectly noticed via the `create_sibling` monster workflow not being
able to clone.

This is now directly checked to pinpoint such a failure closer to the
cause.