Add test on duplicate files
It confirms that dataverse is incapable of representing to identical
files under different names.

This is a show-stopper for supporting export mode. Two identical
`.gitattributes` files would make the system trip already. Not to speak
of other placeholder files that are frequently used to let Git keep
otherwise empty files.

Closes #111