BF: Allow for multiple dataverse IDs per annex key
This turns the file ID record for an annex key into a list, since in
export mode there may be several identical files on the remote end.

Implementation treats the record as a list for regular annex mode as
well for consistency. This should actually never happen, but can't see
reason to always fail when we find there are several dataverse IDs on
record.

Closes #233