Let TRANSFEREXPORT_RETRIEVE probe dataverse if not fileid is known
Rational: git-annex would only call this operation, if it has local
knowledge that the remote should have the key. The previous
implementation refused to check with dataverse, if it has something
for the requested path, when no `fileid` was on record locally.

With this change, dataverse is queried as a fallback.

This seems to be in line with other operations: "use local knowledge
for speed, but not as gospel".