Let `(un)mangle_path()` operate correctly with `PurePosixPath`
Before it would utilize platform paths and also return them. This
created downstream issues, but was also conceptually wrong.

For example `is_dir()` performs a filesystem-based test, but there will
never be the a real path for a dataverse remote path.

The handling of the "directory-only path" condition can be removed. It
cannot occur, because it would not be discriminable.

Closes #259