test: Adjust testremote for expected failure amount
Since Dataverse version 5.0, "storeKey when already present" of git-annex
testremote will fail, as Dataverse forbade replacing files with identical
names and checksums:
https://guides.dataverse.org/en/latest/user/dataset-management.html#duplicate-files
Because we can't exclude any test from the test suite, but wouldn't want
to disable the entire test suite, this change - rather ugly-ly - expects
the check to fail, and tests for the right amount of failures in the
test suite.

Note that it should hopefully be okay to only adjust the tests for this
altered behavior of the remote - a checkpresent prior to transfer
operations should AFAIK prevent upload/replace attempts of identical
files.